chore(refactor notification): Create hook for handling notification messages in reports
