+++ src/Components/SmartComponents/Reports/DownloadCVEsReport.js
@@ -1,8 +1,7 @@
 
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import propTypes from 'prop-types';
 import { useIntl } from 'react-intl';
-import { useDispatch } from 'react-redux';
 import { DownloadButton } from '@redhat-cloud-services/frontend-components-pdf-generator';
 import { getCveListByAccount } from '../../../Helpers/APIHelper';
 import { chunkArray } from '../../../Helpers/MiscHelper';
@@ -13,7 +12,7 @@ import DownloadReport from '../../../Helpers/DownloadReport';
 import { PDF_REPORT_PER_PAGE, STATUS_OPTIONS, BUSINESS_RISK_OPTIONS, impactList } from '../../../Helpers/constants';
 import { Text } from '@react-pdf/renderer';
 import styles from './Common/styles';
-import { addNotification, clearNotifications } from '@redhat-cloud-services/frontend-components-notifications/redux';
+import { useNotification } from '../../../Helpers/Hooks';
 
 const DownloadCVEsReport = ({
     filters,
@@ -27,24 +26,14 @@ const DownloadCVEsReport = ({
     const intl = useIntl();
     const [loading, setLoading] = useState(false);
     const [securityRulesPresent, setSecurityRulesPresent] = useState(false);
-    const dispatch = useDispatch();
-
-    useEffect(() => {
-        return () => {
-            dispatch(clearNotifications());
-        };
-    }, [dispatch]);
+    const addNotification = useNotification();
 
     const otherPagesPDF = data => chunkArray(data, PDF_REPORT_PER_PAGE.otherPages, true);
 
     const cvesFetch = async () => {
         setLoading(true);
 
-        dispatch(addNotification({
-            dismissable: true,
-            variant: 'info',
-            title: intl.formatMessage(messages.customReportNotificationCve)
-        }));
+        addNotification({ msg: messages.customReportNotificationCve });
 
         let { data, meta } = await getCveListByAccount({
             ...params,
