+++ src/Components/SmartComponents/Reports/DownloadExecutive.js
@@ -9,6 +9,8 @@ import { Fragment } from 'react';
 import { useDispatch } from 'react-redux';
 import { Text } from '@react-pdf/renderer';
 import { addNotification, clearNotifications } from '@redhat-cloud-services/frontend-components-notifications/redux';
+import { useNotification } from '../../../Helpers/Hooks';
+
 import styles from './Common/styles';
 
 const DownloadExecutive = () => {
@@ -16,17 +18,14 @@ const DownloadExecutive = () => {
     const [isLoading, setLoading] = useState(false);
     const [renderPDF, setRenderPDF] = useState(false);
     const [securityRulesPresent, setSecurityRulesPresent] = useState(false);
+    const addNotification = useNotification();
 
     const dispatch = useDispatch();
 
     const dataFetch = async () => {
         setLoading(true);
 
-        dispatch(addNotification({
-            dismissable: true,
-            variant: 'info',
-            title: intl.formatMessage(messages.downloadExecutiveNotification)
-        }));
+        addNotification({ msg: messages.downloadExecutiveNotification });
 
         const data = await getExecutiveReport();
         const report = buildExecReport({ data, intl });
