+++ src/Components/SmartComponents/Reports/DownloadSystemsReport.js
@@ -1,8 +1,7 @@
 
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import propTypes from 'prop-types';
 import { useIntl } from 'react-intl';
-import { useDispatch } from 'react-redux';
 import { DownloadButton } from '@redhat-cloud-services/frontend-components-pdf-generator';
 import { getSystems } from '../../../Helpers/APIHelper';
 import { chunkArray } from '../../../Helpers/MiscHelper';
@@ -11,7 +10,7 @@ import firstPagePDF from './Common/firstPagePDF';
 import tablePage from './Common/tablePage';
 import DownloadReport from '../../../Helpers/DownloadReport';
 import { PDF_REPORT_PER_PAGE } from '../../../Helpers/constants';
-import { addNotification, clearNotifications } from '@redhat-cloud-services/frontend-components-notifications/redux';
+import { useNotification } from '../../../Helpers/Hooks';
 
 const DownloadSystemsReport = ({
     filters,
@@ -23,24 +22,14 @@ const DownloadSystemsReport = ({
 }) => {
     const intl = useIntl();
     const [loading, setLoading] = useState(false);
-    const dispatch = useDispatch();
-
-    useEffect(() => {
-        return () => {
-            dispatch(clearNotifications());
-        };
-    }, [dispatch]);
+    const addNotification = useNotification();
 
     const otherPagesPDF = data => chunkArray(data, PDF_REPORT_PER_PAGE.systemsReportRestPages);
 
     const systemsFetch = async () => {
         setLoading(true);
 
-        dispatch(addNotification({
-            dismissable: true,
-            variant: 'info',
-            title: intl.formatMessage(messages.customReportNotificationSystem)
-        }));
+        addNotification({ msg: messages.customReportNotificationSystem });
 
         let { data, meta } = await getSystems({
             ...params,
