+++ src/Helpers/Hooks.js
@@ -1,0 +1,36 @@
+import { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import { useIntl } from 'react-intl';
+import {
+    addNotification as addNotificationFEC,
+    clearNotifications
+} from '@redhat-cloud-services/frontend-components-notifications/redux';
+
+export const useNotification = (config) => {
+    const dispatch = useDispatch();
+    const intl = useIntl();
+
+    const globalConfig = {
+        variant: 'info',
+        dismissable: true,
+        ...config
+    };
+
+    useEffect(() => {
+        return () => {
+            dispatch(clearNotifications());
+        };
+    }, [dispatch]);
+
+    const addNotification = ({ msg, ...config }) =>
+        dispatch(
+            addNotificationFEC({
+                title: intl.formatMessage(msg),
+                ...globalConfig,
+                ...config
+            })
+        );
+
+    return addNotification;
+};
+
